IDL to Python Generation
Originally I was going to generate Python classes "on the fly", but I
got caught up in the idea that users should be able to look at the
generated code and easily understand it, sorta like they should
theoretically be able to do with C++ code (although tao_idl code looks
terrible in practice).

The problem is that IDL/C models of including and forward declaration
don't perfectly fit with the Python import system. This wouldn't be a
problem but want I there to be default values and that doesn't seem like
it's easy to do. The result was circular import issues that I couldn't
get around.

For now I'm continuing the dynamic type route, at least until I reach
the 0.1, then I would like to reassess the feasibility of generating
normal Python code.